sign up done

+++ server/helper.js
@@ -92,8 +92,8 @@ export default class Helper {
     const email = req.body.email;
     const otp = req.body.otp;
     console.log(email, typeof otp);
-    const user = await adminDB.findOne(adminDB.otp, { email: email, otp: otp });
-    if (user) {
+    const user = await adminDB.findOne(adminDB.otp, { email: email });
+    if (user && user.otp == otp) {
       // if OTP matches then send success response
       await adminDB.deleteOne(adminDB.otp, { email: email });
       res.status(200).json({ message: "OTP Verified" });
@@ -105,7 +105,6 @@ export default class Helper {
   static sendOTP = async (req, res) => {
     // extract data from request body
     console.log(req.body);
-    const email = req.body.email;
     // check if user already exists
     const user = await adminDB.findOne(adminDB.users, { email: email });
     if (user) {
@@ -114,6 +113,7 @@ export default class Helper {
       // if user does not exist then register the user
       // send the random OTP to the user
       const otp = Math.floor(Math.random() * 1000000);
+      console.log("OTP: " + otp);
       await adminMail.sendOTP(email, otp);
       const checkOTP = await adminDB.findOne(adminDB.otp, { email: email });
       if (checkOTP) {
